AM-968: Added a minor code refactor.
